+++ CODE/Maths_python/requirement.txt
@@ -1,0 +1,1 @@
+Only library used : numpy
