+++ CODE/PasswordGenerator_python/requirement.txt
@@ -1,0 +1,4 @@
+Library used :
+
+1. random
+2. string
